requirements.txt
schedule
requests
pillow
pytz
python-dotenv
google-api-python-client